% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_utils.R
\name{e2e_collect_and_store}
\alias{e2e_collect_and_store}
\title{End-to-end pipeline}
\usage{
e2e_collect_and_store(
  categories = c("cs.CR", "cs.NI"),
  max_results = 200,
  out = "data-raw",
  strict_mode = TRUE,
  categorize = TRUE,
  category_mode = c("primary", "multi"),
  use_duckdb = FALSE,
  duckdb_path = "inst/data/arxiv.duckdb",
  verbose = TRUE
)
}
\description{
End-to-end pipeline
}
