% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_data.R
\name{fetch_arxiv_data}
\alias{fetch_arxiv_data}
\title{Fetch data from arXiv API}
\usage{
fetch_arxiv_data(categories = "cs.CR", max_results = 10, verbose = TRUE)
}
\description{
Fetch data from arXiv API
}
