% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{init_duckdb}
\alias{init_duckdb}
\title{Initialize DuckDB connection}
\usage{
init_duckdb(db_path = "data/arxiv.duckdb")
}
\arguments{
\item{db_path}{path to DuckDB file}
}
\value{
duckdb connection
}
\description{
Initialize DuckDB connection
}
