% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage_utils.R
\name{init_duckdb_store}
\alias{init_duckdb_store}
\title{Initialize DuckDB and write tables}
\usage{
init_duckdb_store(tables, db_path = "inst/data/arxiv.duckdb")
}
\description{
Initialize DuckDB and write tables
}
