% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{read_from_duckdb}
\alias{read_from_duckdb}
\title{Read data from DuckDB}
\usage{
read_from_duckdb(con, query = NULL)
}
\arguments{
\item{con}{DuckDB connection}

\item{query}{SQL query or NULL}
}
\value{
tibble
}
\description{
Read data from DuckDB
}
