% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{run_collection_pipeline}
\alias{run_collection_pipeline}
\title{Run full collection and storage pipeline}
\usage{
run_collection_pipeline(
  categorize = TRUE,
  category_mode = c("primary", "multi")
)
}
\description{
Run full collection and storage pipeline
}
