% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_data.R
\name{save_collected_data}
\alias{save_collected_data}
\title{Save collected data to a file
#' @export}
\usage{
save_collected_data(data, file_path, compress = TRUE)
}
\description{
Save collected data to a file
#' @export
}
