% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_utils.R
\name{write_to_duckdb}
\alias{write_to_duckdb}
\title{Write normalized data to DuckDB}
\usage{
write_to_duckdb(normalized_data, con)
}
\arguments{
\item{normalized_data}{list from normalize_arxiv_data}

\item{con}{DuckDB connection}
}
\description{
Write normalized data to DuckDB
}
