% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect_data.R
\name{filter_cybersecurity}
\alias{filter_cybersecurity}
\title{Filter cybersecurity publications 
#' @export}
\usage{
filter_cybersecurity(data, strict_mode = FALSE)
}
\description{
Filter cybersecurity publications 
#' @export
}
